+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ### Added
 
+- MixedPool to `allocator_benchmarks.cpp`
+
 ### Changed
 
 ### Removed